Move componente Carrosel, adiciona id do livro na rota
